MOHAWK: Move the external commands to their respective stacks
