+++ engines/mohawk/module.mk
@@ -54,7 +54,6 @@ ifdef ENABLE_RIVEN
 MODULE_OBJS += \
 	riven.o \
 	riven_card.o \
-	riven_external.o \
 	riven_graphics.o \
 	riven_saveload.o \
 	riven_scripts.o \
