+++ engines/mohawk/riven_stacks/pspit.h
@@ -28,9 +28,23 @@
 namespace Mohawk {
 namespace RivenStacks {
 
+/**
+ * Prison Island
+ */
 class PSpit : public DomeSpit {
 public:
 	PSpit(MohawkEngine_Riven *vm);
+
+	// External commands - Prison Elevator
+	void xpisland990_elevcombo(uint16 argc, uint16 *argv);	// Param1: button
+
+	// External commands - Dome
+	void xpscpbtn(uint16 argc, uint16 *argv);
+	void xpisland290_domecheck(uint16 argc, uint16 *argv);
+	void xpisland25_opencard(uint16 argc, uint16 *argv);
+	void xpisland25_resetsliders(uint16 argc, uint16 *argv);
+	void xpisland25_slidermd(uint16 argc, uint16 *argv);
+	void xpisland25_slidermw(uint16 argc, uint16 *argv);
 
 };
